Move helper function around b/c it will be used in another module
